Update scripts for new updater

+++ lmodroid/android-builder/publish.sh
@@ -8,23 +8,24 @@ cd $ROOTDIR
 BUILT_ROM_PATH=`ls $ROOTDIR/out/target/product/${DEVICE}/LMODroid-*.zip`
 if [[ $? -eq 0 ]]; then
     echo "[+] Built Rom: ${BUILT_ROM_PATH}"
-    ROM_BASENAME=`basename ${BUILT_ROM_PATH} ".zip"`
-    RECOVERY_NAME=`echo "${ROM_BASENAME}.img" | sed "s/-${DEVICE}/-recovery-${DEVICE}/"`
-
-    RECOVERYIMG=boot.img
-    if [[ -f $ROOTDIR/out/target/product/$DEVICE/recovery.img ]]; then
-        RECOVERYIMG=recovery.img
-    fi
+    DATETIME=`echo ${BUILT_ROM_PATH%.zip} | cut -f 3 -d '-'`
+    BUILT_IMAGES="boot recovery dtbo vendor_boot super_empty"
+    TARGETFILES=`find out/target/product/${DEVICE}/obj/PACKAGING/target_files_intermediates/ -maxdepth 1 -type d  | grep target_files-`
 
     echo "[+] Upload ROM to OTA server."
+    ssh -p 40048 -o StrictHostKeyChecking=no root@get.libremobileos.com mkdir -p /root/builds/full/${DEVICE}/${DATETIME}
     rsync -avz "--rsh=ssh -p 40048 -o StrictHostKeyChecking=no" \
         ${BUILT_ROM_PATH} \
-        root@get.libremobileos.com:/root/builds/full/
+        root@get.libremobileos.com:/root/builds/full/${DEVICE}/${DATETIME}
 
-    echo "[+] Upload Recovery to OTA server."
-    rsync -avz "--rsh=ssh -p 40048 -o StrictHostKeyChecking=no" \
-        $ROOTDIR/out/target/product/$DEVICE/${RECOVERYIMG} \
-        root@get.libremobileos.com:/root/builds/recovery/${RECOVERY_NAME}
+    for IMAGE in $BUILT_IMAGES; do
+        if [ -f $TARGETFILES/IMAGES/$IMAGE.img ]; then
+            echo "[+] Upload $IMAGE to OTA server."
+            rsync -avz "--rsh=ssh -p 40048 -o StrictHostKeyChecking=no" \
+                $TARGETFILES/IMAGES/$IMAGE.img \
+                root@get.libremobileos.com:/root/builds/full/${DEVICE}/${DATETIME}
+        fi
+    done
 else
     exit 1
 fi
